add fullscreen mode (escape button)
