+++ include/scene.hpp
@@ -65,23 +65,6 @@ public:
     void handleInput(sf::Event &event) override;
 };
 
-class LevelScene : public Scene {
-private:
-    std::string ldtk_filename;
-    Player player;
-    LevelsStorage storage;
-    int levelNumber;
-    Level level;  // добавила класс уровня
-
-public:
-    LevelScene() = default;
-    LevelScene(std::string ldtk_filename);
-    LevelScene(int levelN);
-    void update(sf::Time &dTime) override;
-    void draw(sf::RenderWindow &window) override;
-    void handleInput(sf::Event &event) override;
-};
-
 }  // namespace super_hse
 
 #endif  // SCENE_HPP_
