+++ src/game.cpp
@@ -2,10 +2,16 @@
 #include <memory>
 #include "scene.hpp"
 
+
+#include <iostream>
+
 namespace super_hse {
 
 int Game::windowWidth = 960;
 int Game::windowHeight = 640;
+
+int Game::defaultWindowWidth = 960;
+int Game::defaultWindowHeight = 640;
 
 void Game::run() {
     SceneManager::changeScene(std::make_unique<MainMenuScene>());
@@ -20,6 +26,19 @@ void Game::run() {
             }
             if (event.type == sf::Event::KeyPressed && event.key.code == sf::Keyboard::Escape){
                 window.close();
+
+                if (isFullScreen) {
+                    windowWidth = defaultWindowWidth;
+                    windowHeight = defaultWindowHeight;
+                    window.create(sf::VideoMode(windowWidth, windowHeight, 32), "Super HSE", sf::Style::Default);
+
+                } else {
+                    window.create(sf::VideoMode::getDesktopMode(), "Super HSE", sf::Style::Fullscreen);
+                    windowWidth = window.getSize().x;
+                    windowHeight = window.getSize().y;
+                }
+
+                isFullScreen = !isFullScreen;
             }
             SceneManager::handleInput(event);
         }
